+++ src/html/text.ts
@@ -1,8 +1,10 @@
 import { hasFeature } from '@/config/feature';
+import { getDevice } from '@delegates/retrievers/device';
+import { getEntity } from '@delegates/retrievers/entity';
+import { getState } from '@delegates/retrievers/state';
 import type { HomeAssistant } from '@hass/types';
 import type { Config } from '@type/config';
 import { html, nothing, type TemplateResult } from 'lit';
-import { getDevice, getEntity, getState } from '../helpers';
 
 /**
  * Gets the climate label combining temperature and humidity when available
