+++ test/cards/index.spec.ts
@@ -1,0 +1,5 @@
+import editorSpec from './editor.spec';
+
+describe('cards', () => {
+  editorSpec();
+});
